Added delete function and documentation
