+++ Java/SQLConnectivity/description.md
@@ -1,0 +1,3 @@
+##Java class to demonstrate CRUD operation in java 
+DBConnection class has methods to create, insert, update, delete and show records in Student table.
+client has the main method which executes methods of DBConnection  class
